% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.summary.R
\name{input.summary}
\alias{input.summary}
\title{Summary of the input paramters used by \code{PEG} and \code{PTL} functions}
\usage{
input.summary(data)
}
\arguments{
\item{data}{dataframe from \code{PEG} and \code{PTL} functions input datas. Other inputs will give an error.}
}
\value{
A dataframe with parameters and dispersion statistics in raw and, species (SP) and safety factor (Fs) or objective factor (Fobj) in column :
In columns :
\itemize{
  \item \code{SP} the species
}
In rows :
\itemize{
  \item \code{Fs} safety factor
  \item \code{Fobj} objective factor
  \item \code{mean.surv} arithmetic mean of annual survival rate for the adult age class
  \item \code{med.surv} median of annual survival rate for the adult age class
  \item \code{se.surv} standard error of annual survival rate for the adult age class
  \item \code{surv.CI95.low } lower bound of the 95 \% confidence interval for annual survival rate
  \item \code{surv.CI95.up} upper bound of the 95 \% confidence interval for annual survival rate
  \item \code{mean.alpha} arithmetic mean of age at first breeding
  \item \code{med.alpha} median of age at first breeding
  \item \code{se.alpha} standard error of annual survival rate for the adult age class
  \item \code{alpha.CI95.low } lower bound of the 95 \% confidence interval for age at first breeding
  \item \code{alpha.CI95.up} upper bound of the 95 \% confidence interval for age at first breeding
  \item \code{mean.Rmax} arithmetic mean of age at first breeding
  \item \code{med.Rmax} median of the maximum annual growth rate
  \item \code{se.Rmax} standard error of the maximum annual growth rate
  \item \code{Rmax.CI95.low } lower bound of the 95 \% confidence interval of the maximum annual growth rate
  \item \code{Rmax.CI95.up} upper bound of the 95 \% confidence interval of the maximum annual growth rate
}
}
\description{
The function returns descriptive statistics (mean, median, se and 95 \% confidence intervals) for the input parameters used by PEG and PTL functions. These parameters are : annual survival rate of adults (surv), age at first breeding (alpha), maximum multiplicative growth rate (lambdaMax) and maximum annual population growth rate (Rmax). Descriptive statistics are drawn from the n simulations called within the \code{PEG} and \code{PTL} functions. When one or more parameters are fixed by the user, the corresponding value is reported for mean and median, without any further measure of dispersion (i.e. se and 95 \% CI are equal to zero). Statistics are reported in columns.
}
\examples{
df <- PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

input.summary(df)

}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
