% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.summary.R
\name{output.summary}
\alias{output.summary}
\title{Summary of the output parameters from the \code{PEG} and the \code{PTL} functions}
\usage{
output.summary(data)
}
\arguments{
\item{data}{dataframe generated by the \code{PEG} function or the the \code{PTL} function detailing calculations for each of the n iterations. Other input will give an error.}
}
\value{
A dataframe containing the following statistics calculated over the n iterations:
In columns :
\itemize{
  \item \code{SP} the species
}
In rows :
\itemize{
  \item \code{Fs} safety factor or \code{Fobj} objective factor
  \item \code{med.Rmax} median of the maximum annual growth rate
  \item \code{mean.Rmax} arithmetic mean of the maximum annual growth rate
  \item \code{se.Rmax} standard error of the maximum annual growth rate
  \item \code{Rmax.CI95.low} lower bound of the 95 \% confidence interval of the maximum annual growth rate
  \item \code{Rmax.CI95.up} upper bound of the 95 \% confidence interval of the maximum annual growth rate
  \item \code{med.PEG} median of the potential excess growth or \code{med.PTL} median of the maximum theoretical harvest
  \item \code{mean.PEG} arithmetic mean of the potential excess growth or \code{mean.PTL} arithmetic mean of the maximum theoretical harvest
  \item \code{se.PEG} standard error of the potential excess growth or \code{se.PTL} standard error of the maximum theoretical harvest
  \item \code{PEG.CI95.low} lower bound of the 95 \% confidence interval of the potential excess growth or \code{PTL.CI95.low} lower bound of the 95 \% confidence interval of the maximum theoretical harvest
  \item \code{PEG.CI95.up} upper bound of the 95 \% confidence interval of the potential excess growth or \code{PTL.CI95.up} upper bound of the 95 \% confidence interval of the maximum theoretical harvest
  \item \code{med.SHI} median of the sustainability harvest index
  \item \code{mean.SHI} arithmetic mean of the sustainability harvest index
  \item \code{se.SHI} standard error of the sustainability harvest index
  \item \code{SHI.CI95.low} lower bound of the 95 \% confidence interval of the sustainability harvest index
  \item \code{SHI.CI95.up} upper bound of the 95 \% confidence interval of the sustainability harvest index
  \item \code{unsustain.harvest} frequency (1 = 100\%) of the n iterations  with sign of overharvesting
}
}
\description{
The function returns descriptive statistics (mean, median, se and 95 \% confidence intervals) for the output parameters used by PEG and PTL functions. These parameters are : potential excess growth (PEG) or prescribed take level (PTL), sustainable harvest index (SHI), the maximum annual population growth rate (Rmax) and the number of iterations for which the observed harvest exceeds the sustainable harvest index (unsustain.harvest). Descriptive statistics are drawn from the n simulations called within the \code{PEG} and \code{PTL} functions. When one or more parameters are fixed by the user, the corresponding value is reported for mean and median, without any further measure of dispersion (i.e. se and 95 \% CI are equal to zero). Statistics are reported in columns.
}
\examples{
df <- PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

output.summary(df)

}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}
}
