% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEG.R
\name{PEG}
\alias{PEG}
\title{Potential Excess Growth estimation}
\usage{
PEG(
  full.option = FALSE,
  Nsim = 1,
  NSp = NULL,
  Rmax.fixed = NULL,
  Rmax.norm = FALSE,
  mean.Rmax = NULL,
  sd.Rmax = NULL,
  lambdaMax.fixed = NULL,
  lambdaMax.norm = FALSE,
  mean.lambdaMax = NULL,
  sd.lambdaMax = NULL,
  living.rate = NULL,
  surv.fixed = NULL,
  surv.norm = FALSE,
  mean.surv = NULL,
  sd.surv = NULL,
  mass.fixed = NULL,
  mass.norm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  type.p = NULL,
  type.e = NULL,
  alpha.fixed = NULL,
  alpha.unif = FALSE,
  min.alpha = NULL,
  max.alpha = NULL,
  alpha.norm = FALSE,
  mean.alpha = NULL,
  sd.alpha = NULL,
  Fs = NULL,
  harvest.fixed = NULL,
  harvest.unif = FALSE,
  min.harvest = NULL,
  max.harvest = NULL,
  harvest.norm = FALSE,
  mean.harvest = NULL,
  sd.harvest = NULL,
  pop.fixed = NULL,
  pop.unif = FALSE,
  min.pop = NULL,
  max.pop = NULL,
  pop.norm = FALSE,
  mean.pop = NULL,
  sd.pop = NULL
)
}
\arguments{
\item{full.option}{boolean, default \code{FALSE}. If \code{TRUE}, provides all parameter values necessary for calculations.}

\item{Nsim}{number, default 1 ;  number of simulations.}

\item{NSp}{number ; number of species for which calculations are done.}

\item{Rmax.fixed}{number ; fixed estimate of maximum annual growth rate provided by the user. To be used instead of \code{mean.Rmax} when \code{Rmax.norm = FALSE}.}

\item{Rmax.norm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates Rmax for a normal distribution with mu = \code{mean.Rmax} and standard deviation = \code{sd.Rmax}.}

\item{mean.Rmax}{number ; arithmetic mean of maximum annual growth rate.}

\item{sd.Rmax}{number ; standard deviation of maximum annual growth rate.}

\item{lambdaMax.fixed}{number ; fixed estimate of maximum multiplicative growth rate provided by the user and used to calculate Rmax as lamndaMax-1.}

\item{lambdaMax.norm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates maximum multiplicative growth rate from a normal distribution with mu = \code{mean.lambdaMax} and standard deviation = \code{sd.lambdaMax}.}

\item{mean.lambdaMax}{number ; arithmetic mean of maximum multiplicative growth rate. To by used only if \code{lamndaMax.norm = TRUE}.}

\item{sd.lambdaMax}{number ; standard deviation of maximum multiplicative growth rate. To be used only if \code{lambdaMax.norm = TRUE}.}

\item{living.rate}{character string ; living rate : either "\code{long}" or "\code{short}". No other character string allowed.}

\item{surv.fixed}{number ; fixed estimate of adult annual survival rate provided by the user. To be used instead of \code{mean.surv} when \code{surv.norm = FALSE}.}

\item{surv.norm}{boolean, default FALSE. If \code{TRUE}, simulates adult annual survival rate from a normal distribution with mu = \code{mean.surv} and standard deviation = \code{sd.surv} (see below).}

\item{mean.surv}{number ; arithmetic mean of adult annual survival rate. Only used if \code{surv.norm = TRUE}.}

\item{sd.surv}{number ; standard deviation of adult annual survival rate. Only used if \code{surv.norm = TRUE}.}

\item{mass.fixed}{number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate from body mass.}

\item{mass.norm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.norm = TRUE}.}

\item{sd.mass}{number ; standard deviation of body mass. Only used if \code{mass.norm = TRUE}.}

\item{type.p}{character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{alpha.fixed}{number ; fixed estimate of age at first breeding (in years). To be used when \code{alpha.unif = FALSE} and \code{alpha.norm = FALSE}.}

\item{alpha.unif}{boolean, default \code{FALSE}. Simulates age at first breeding from a uniform distribution bounded by \code{min.alpha} and \code{max.alpha}.}

\item{min.alpha}{number ; minimum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{max.alpha}{number ; maximum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{alpha.norm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates age at first breeding from a normal distribution with mu = \code{mean.alpha} and standard deviation = \code{sd.alpha}.}

\item{mean.alpha}{number ; arithmetic mean of age at first breeding. To be used only used if \code{alpha.norm = TRUE}.}

\item{sd.alpha}{number ; age at first breeding standard deviation. Only used if \code{alpha.norm = TRUE}.}

\item{Fs}{number ; safety factor, see 'Details'.}

\item{harvest.fixed}{number ; number of harvested individuals.}

\item{harvest.unif}{boolean ; default \code{FALSE}. If \code{TRUE}, simulates harvest level from a uniform distribution bounded by \code{min.harvest} and \code{max.harvest}.}

\item{min.harvest}{number ; minimum number of harvested individuals. To be used only if \code{harvest.unif = TRUE}.}

\item{max.harvest}{number ; maximum number of harvested individuals. To be used only if \code{harvest.unif = TRUE}.}

\item{harvest.norm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates harvest level from a normal distribution with mu =  \code{mean.harvest} and standard deviation = \code{sd.harvest}.}

\item{mean.harvest}{number ; arithmetic mean of the number of harvested individuals. To be used only if \code{harvest.norm = TRUE}.}

\item{sd.harvest}{number ; standard deviation of the number of harvested individuals. To be used only if \code{harvest.norm = TRUE}.}

\item{pop.fixed}{number ; fixed estimate of population size (in number of individuals). To be used when \code{pop.unif = FALSE} and \code{pop.norm = FALSE}.}

\item{pop.unif}{boolean, default \code{FALSE}. If \code{TRUE}, simulates a population size from a uniform distribution bounded by \code{min.pop} and \code{max.pop}.}

\item{min.pop}{number ; minimum population size. To be used only if \code{pop.unif = TRUE}.}

\item{max.pop}{number ; maximum population size. To be used only if \code{pop.unif = TRUE}.}

\item{pop.norm}{boolean, default \code{FALSE}.  If \code{TRUE}, simulates a population size from a normal distribution with mu = \code{mean.pop} and standard deviation = \code{sd.pop}.}

\item{mean.pop}{number ; arithmetic mean of population size. To be used only if \code{pop.norm = TRUE}.}

\item{sd.pop}{number ; standard deviation of population size. To be used only if \code{pop.norm = TRUE}.}
}
\value{
A dataframe with the following information for each of the n simulations :
If the argument \code{full.option} is \code{FALSE}, it provides a simple dataframe :
\itemize{
  \item \code{SP} the identification number assigned to the studied species.
  \item \code{Sim} the index of the simulation.
  \item \code{Fs} the value of the safety factor.
  \item \code{Rmax} the maximal annual recruitment rate (lambda max – 1).
  \item \code{popsize} the population size (in number of individuals).
  \item \code{harvest}  the number of individuals harvested each year.
  \item \code{PEG} the calculated value of the Potential excess growth (in number of individuals).
  \item \code{SHI} the calculated value of the Sustainability harvest index (from 0 to +∞) (if > 1, harvest exceeds PEG).
}
If the argument \code{full.option} is \code{TRUE}, it can provide the following informations in addition to the these within the simple dataframe :
\itemize{
  \item \code{e} a value extracted from a normal distribution (0, 0.087). This parameter appears when the adult annual survival rate is estimated from the body mass.
  \item \code{p} a value extracted from a beta distribution (3.34,101.24). This parameter appears when the adult annual survival rate is estimated from the body mass.
}
}
\description{
\code{PEG} is a function used to assess whether a known level of harvesting exceeds the biological potential of renewal of a population, see 'Details'.
}
\details{
The function calculates the Potential Excess Growth (PEG) and Sustainability Harvest Index (SHI), in absence of others mortality source.


PEG is expressed as a number of individuals and corresponds to the maximum number that can be removed annually from a population
given its maximum annual growth rate (Rmax). PEG is calculated as the product of the safety factor (Fs), Rmax, and the population size (popsize).


Fs (\code{Fs} argument) is a positive number less than 1 that takes into account several aspects such as the uncertainties
associated with estimates of population size and maximum annual population growth rate as well as the shape and magnitude of the
dependence density phenomena involved. This is a precaution that takes only a fraction of potential growth in
excess.


Rmax is the maximum annual recruitment rate estimation. This is equivalent to the maximum annual population
growth rate, lambdaMax, under optimal growth conditions. Rmax and lambdaMax can be provided by either a point estimates (\code{Rmax.fixed}, \code{lambdaMax.fixed}) or a normal distribution (\code{Rmax.norm = TRUE}
with \code{mean.Rmax} and \code{sd.Rmax}, \code{lambdaMax = TRUE} with \code{mean.lambdaMax} and \code{sd.lambdaMax}). When Rmax (or lambdaMax) is unknown from the user,
the value is calculated using the demographic invariants
approach (Niel & Lebreton 2005). This approach calculates \eqn{Rmax = lambda max – 1} on the basis of 2 parameters:
adult survival rate and age at first breeding. When adult survival and age at first reproduction are known
from the user, they can be provided by either point estimates (\code{surv.fixed}, \code{alpha.fixed} arguments) or distributions (\code{surv.norm = TRUE} with \code{mean.surv} and \code{sd.surv},
\code{alpha.unif = TRUE} with \code{min.alpha} and \code{max.alpha}, \code{alpha.norm = TRUE} with \code{mean.alpha} and \code{sd.alpha} arguments). Also,
survival rate can be estimated from body mass following the relationship described in Johnson et al. (2012) : \eqn{p^(1/(exp(3.22+0.24*log(body mass)+e)-1))}.
In this case, body mass can be provided as either a point estimate (\code{mass.fixed})
or a normal distribution (\code{mass.norm = TRUE}, \code{mean.mass} and \code{sd.mass} arguments). The parameter p can be provided by either from a beta distribution
B(3.34, 101.24) (\code{type.p = "random"}) or a point estimate equal to 0.032 (i.e. the mean of the beta distribution is given by a/(a + b) with a = 3.34 and b = 101.24, Boitani & Fuller 2000) (\code{type.p = "determinist"})
The parameter e can also be either null (\code{type.e = "determinist"}) or residuals from a normal distribution N(0, 0.087) (\code{type.e = "random"}).


In all cases,  the type of living rate (\code{living.rate} argument), i.e. long-lived species (\code{living.rate = "long"}) or short-lived species (\code{living.rate = "short"}) must be quoted. Long-lived species,
unlike short-lived species, are those that do not have a strategy to maximize reproduction, i.e. those with high adult survival, low brood size,
and high first-generation age. This argument is used to select the right formula from Niel & Lebreton (2005) for the calculation of lambdaMax
and, therefater Rmax.


The population size, as the previous terms, can also be provided by either a point estimates (\code{pop.fixed}) or distributions (\code{pop.unif = TRUE}
with \code{min.pop} and \code{max.pop} or \code{pop.norm = TRUE} with \code{mean.pop} and \code{sd.pop}).


The Sustainability Harvest Index or SHI is the harvest level expressed as a proportion relative to PEG: \eqn{harvest level/PEG}. The harvest level can
be provided by a point estimates (\code{harvest.fixed}) or distributions (\code{harvest.unif = TRUE}
with \code{min.harvest} and \code{max.harvest} or \code{harvest.norm = TRUE} with \code{mean.harvest} and \code{sd.harvest}). In natura, it is
unlikely that a population is at maximum annual growth rate (Rmax). Consequently, when harvest exceeds PEG, this suggests an
overexploitation regime.


The function PEG provides a great flexibility of use according the known arguments and analysis that the user wants realize.
Fixed demographic parameters or different distribution types can be integrated in the function. This later approach enables uncertainty
to be taken into account.


For more details about the use of the function, see 'README' documentation.
}
\examples{
PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

}
\references{
Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.
Runge, M. C., Sauer, J. R., Avery, M. L., Blackwell, B. F., & Koneff, M. D. (2009). Assessing allowable take of migratory birds. The Journal of Wildlife Management, 73(4), 556-565.

Boitani, L., & Fuller, T. (Eds.). (2000). Research techniques in animal ecology: controversies and consequences. Columbia university press.
}
\seealso{
\code{\link{input.summary}}, \code{\link{output.summary}}.
}
