% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL.R
\name{PTL}
\alias{PTL}
\title{Prescribed Take Level estimation}
\usage{
PTL(
  full.option = FALSE,
  Nsim = 1,
  NSp = NULL,
  Rmax.fixed = NULL,
  Rmax.lognorm = FALSE,
  mean.Rmax = NULL,
  sd.Rmax = NULL,
  lambdaMax.fixed = NULL,
  lambdaMax.lognorm = FALSE,
  mean.lambdaMax = NULL,
  sd.lambdaMax = NULL,
  living.rate = NULL,
  surv.fixed = NULL,
  surv.beta = FALSE,
  mean.surv = NULL,
  sd.surv = NULL,
  mass.fixed = NULL,
  mass.lognorm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  type.p = NULL,
  type.e = NULL,
  surv.j = 1,
  alpha.fixed = NULL,
  alpha.unif = FALSE,
  min.alpha = NULL,
  max.alpha = NULL,
  alpha.lognorm = FALSE,
  mean.alpha = NULL,
  sd.alpha = NULL,
  Fobj = NULL,
  harvest.fixed = NULL,
  harvest.unif = FALSE,
  min.harvest = NULL,
  max.harvest = NULL,
  harvest.lognorm = FALSE,
  mean.harvest = NULL,
  sd.harvest = NULL,
  pop.fixed = NULL,
  pop.unif = FALSE,
  min.pop = NULL,
  max.pop = NULL,
  pop.lognorm = FALSE,
  mean.pop = NULL,
  sd.pop = NULL,
  theta.fixed = NULL,
  estim.theta = NULL
)
}
\arguments{
\item{full.option}{boolean, default \code{FALSE}. If \code{TRUE}, provides all parameter values necessary for calculations.}

\item{Nsim}{number, default 1 ;  number of simulations.}

\item{NSp}{number ; number of species for which calculations are done.}

\item{Rmax.fixed}{number ; fixed estimate of maximum annual recruitment rate provided by the user. To be used instead of \code{mean.Rmax} when \code{Rmax.lognorm = FALSE}.}

\item{Rmax.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates Rmax from a log-normal distribution with mu = \code{mean.Rmax} and standard deviation = \code{sd.Rmax}.}

\item{mean.Rmax}{number ; arithmetic mean of maximum annual recruitment rate.}

\item{sd.Rmax}{number ; standard deviation of maximum annual recruitment rate.}

\item{lambdaMax.fixed}{number ; fixed estimate of maximum annual multiplicative growth rate provided by the user and used to calculate Rmax as \eqn{lamdaMax-1}.}

\item{lambdaMax.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates maximum annual multiplicative growth rate from a log-normal distribution with mu = \code{mean.lambdaMax} and standard deviation = \code{sd.lambdaMax}.}

\item{mean.lambdaMax}{number ; arithmetic mean of maximum annual multiplicative growth rate. To be used only if \code{lamndaMax.lognorm = TRUE}.}

\item{sd.lambdaMax}{number ; standard deviation of maximum multiplicative growth rate. To be used only if \code{lambdaMax.lognorm = TRUE}.}

\item{living.rate}{character string ; living rate : either \code{"long"} or \code{"short"}, see 'Details'. No ohter arguments allowed.}

\item{surv.fixed}{number ; fixed estimate of (annual) adult survival rate provided by the user. To be used instead of \code{mean.surv} when \code{surv.beta = FALSE}.}

\item{surv.beta}{boolean, default \code{FALSE}. If \code{TRUE}, simulates (annual) adult survival rate from a beta distribution with mu = \code{mean.surv} and standard deviation = \code{sd.surv} (see below).}

\item{mean.surv}{number ; arithmetic mean of (annual) adult survival rate. Only used if \code{surv.beta = TRUE}.}

\item{sd.surv}{number ; standard deviation of adult annual survival rate. Only used if \code{surv.beta = TRUE}.}

\item{mass.fixed}{number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate under ideal conditions from body mass.}

\item{mass.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.lognorm = TRUE}.}

\item{sd.mass}{number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{type.p}{character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{surv.j}{number, default 1 ; point estimate of average annual survival for birds aged 1 to alpha. juvenile annual survival rate The value can be provided by the user . Used to estimate adult survival rate from body mass under certain circumstances (see 'Details').}

\item{alpha.fixed}{number ; number ; fixed estimate of age at first breeding (in years). To be used when \code{alpha.unif = FALSE} and \code{alpha.lognorm = FALSE}.}

\item{alpha.unif}{boolean, default \code{FALSE}. Simulates age at first breeding from a uniform distribution bounded by \code{min.alpha} and \code{max.alpha}.}

\item{min.alpha}{number ; minimum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{max.alpha}{number ; maximum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{alpha.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates age at first breeding from a log-normal distribution with mu = \code{mean.alpha} and standard deviation = \code{sd.alpha}.}

\item{mean.alpha}{number ; arithmetic mean of age at first breeding. To be used only used if \code{alpha.lognorm = TRUE}.}

\item{sd.alpha}{number ; standard deviation of age at first breeding. To be used only if \code{alpha.lognorm = TRUE}.}

\item{Fobj}{number ; objective factor, see 'Details'.}

\item{harvest.fixed}{number ; number of harvested individuals.}

\item{harvest.unif}{boolean ; default \code{FALSE}. If \code{TRUE}, simulates harvest level from a uniform distribution bounded by \code{min.harvest} and \code{max.harvest}.}

\item{min.harvest}{number ; minimum number of harvested individuals. To be used only if \code{harvest.unif = TRUE}.}

\item{max.harvest}{number ; maximum number of individuals harvested. To be used only if \code{harvest.unif = TRUE}.}

\item{harvest.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates harvest level from a log-normal distribution with mu =  \code{mean.harvest} and standard deviation = \code{sd.harvest}.}

\item{mean.harvest}{number ; arithmetic mean of the number of harvested individuals. To be used only if \code{harvest.lognorm = TRUE}.}

\item{sd.harvest}{number ; standard deviation of the number of harvested individuals. To be used only if \code{harvest.lognorm = TRUE}.}

\item{pop.fixed}{number ;  fixed estimate of population size (in number of individuals). To be used when \code{pop.unif = FALSE} and \code{pop.lognorm = FALSE}.}

\item{pop.unif}{boolean, default \code{FALSE}. If \code{TRUE}, simulates a population size from a uniform distribution bounded by \code{min.pop} and \code{max.pop}.}

\item{min.pop}{number ; minimum population size. To be used only if \code{pop.unif = TRUE}.}

\item{max.pop}{number ; maximum population size. To be only used if \code{pop.unif = TRUE}.}

\item{pop.lognorm}{boolean, default \code{FALSE}.  If \code{TRUE}, simulates a population size from a log-normal distribution with mu = \code{mean.pop} and standard deviation = \code{sd.pop}.}

\item{mean.pop}{number ; arithmetic mean of population size. To be used only if \code{pop.lognorm = TRUE}.}

\item{sd.pop}{number ; standard deviation of population size. To be used only if \code{pop.lognorm = TRUE}.}

\item{theta.fixed}{number ; parameter for density-dependance.}

\item{estim.theta}{character string ; default "\code{determinist}". The value of the parameter theta can be either "\code{determinist}" or "\code{random}", see 'Details'. No other arguments allowed.}
}
\value{
A dataframe with the following informations for each the n simulations in addition to the information provided in the function's arguments :
If the argument \code{full.option} is \code{FALSE}, it provides a simple dataframe :
\itemize{
  \item \code{SP} the identification number assigned to the studied species
  \item \code{Sim} the index of the simulation.
  \item \code{Fs} the value of the safety factor.
  \item \code{Rmax} the maximal annual recruitment rate (lambda max – 1).
  \item \code{popsize} the population size (in number of individuals).
  \item \code{harvest}  the number of individuals harvested each year.
  \item \code{PTL} the Prescribed take level expressed in  individuals number.
  \item \code{SHI} Sustainability harvest index (from 0 to +∞).
}
If the argument \code{full.option} is \code{TRUE}, the following informations are added to the previous dataframe :
\itemize{
  \item \code{e} a value drawn from a normal distribution (0, 0.087). This parameter occurs when the (annual) adult survival rate is estimated from body mass.
  \item \code{p} a value drawn from a beta distribution (3.34,101.24). This parameter occurs when the annual survival rate of adults is estimated from body mass.
  \item \code{e.density} a value drawn from a normal distribution (0, 0.942). This parameter occurs when the value of the parameter for density-dependence is unknown and needs to be estimated from Rmax.
}
}
\description{
\code{PTL} is a function used to calculate the prescribed take level of a species, see 'Details'.
}
\details{
Calculates the Prescribed Take Level (PTL) integrating a density-dependence parameter and returns a Sustainable Harvest Index (SHI).

PTL is the maximum number of individuals that can be removed from a population while allowing it to
reach or maintain equilibrium to a certain level of its carrying capacity (Runge \emph{et al.}, 2009). The PTL formula is
\eqn{Fobj.(Rmax.theta)/(1+theta)*population size} for a breeding population (Johnson \emph{ al.}, 2012).


Fobj (\code{Fobj} argument) is a factor that reflects management objectives of the decision-maker. It is related to
the desired long-term population size relative to carrying capacity (Runge et al. 2009).


Rmax is the maximum annual recruitment rate. This is equivalent to \eqn{1-lambdaMax}, where lambdaMax is the maximum annual growth rate
under optimal growth conditions. Rmax and lambdaMax can be provided either as point estimates (\code{Rmax.fixed},\code{lambdaMax.fixed}) or drawn from a log-normal distribution (\code{Rmax.lognorm = TRUE}
with \code{mean.Rmax} and \code{sd.Rmax}, \code{lambdaMax.lognorm = TRUE} with \code{mean.lambdaMax} and \code{sd.lambdaMax}). When Rmax (or lambdaMax) is unknown,
the value can be estimated using the Demographic Invariants Method (DIM) (Niel & Lebreton, 2005) (see \code{rmax} function). This approach calculates \eqn{Rmax = lambdaMax – 1} Dillingham & Fletcher, 2008) on the basis of 2 parameters :
annual survival rate of adults under ideal condition (maximal survival, \code{surv} arguments) and age at first breeding. When adult survival rate (\code{surv} arguments) and age at first reproduction (\code{alpha} arguments) are known,
users can provide corresponding values either as point estimates (\code{surv.fixed}, \code{alpha.fixed} arguments) or as distributions (\code{surv.beta = TRUE} with \code{mean.surv} and \code{sd.surv},
\code{alpha.unif = TRUE} with \code{min.alpha} and \code{max.alpha}, \code{alpha.lognorm = TRUE} with \code{mean.alpha} and \code{sd.alpha} arguments). Alternatively,
adult survival rate can be estimated from body mass following the relationship described in Johnson \emph{et al.} (2012) : \eqn{(p/surv.j)^(1/(exp(3.22+0.24*log(body mass)+e)-alpa))} (see \code{surv} function).
Body mass can be provided either as point estimates (\code{mass.fixed})
or drawn from a log-normal distribution (\code{mass.lognorm = TRUE}, \code{mean.mass} and \code{sd.mass} arguments). The parameter p can be either drawn from a beta distribution
B(3.34, 101.24) (\code{type.p = "random"}) or provided as a point estimate equal to 0.032 (i.e. the mean of the beta distribution is given by a/(a + b) with a = 3.34 and b = 101.24, Boitani & Fuller 2000) (\code{type.p = "determinist"})
The parameter e (residuals) can be either null (set to zero) (\code{type.e = "determinist"}) or residuals from a normal distribution N(0, 0.087) (\code{type.e = "random"}). See equations in Johnson \emph{et al.} (2012).


#' By default, the argument \code{surv.j} is silent and masked to users (the default value for \code{surv.j} is 1, as such \eqn{p/surv.j = p/1 = p} ; see original equation in
Johnson \emph{et al.}, 2012). This is correct for any alpha if, and only if, juvenile survival < adult survival solely for birds aged <1 year.  This is a reasonable
approximation especially in passerines and some other medium-sized bird species. For species with delayed sexual maturity and when juvenile survival < adult
survival for birds aged 1 to alpha, the default \code{surv} function returns biased estimates unless the users specify a value for annual survival for birds aged 1
to alpha. This is done by using the argument \code{surv.j} which is then used to estimate  \eqn{p = p/surv.j}.


In all cases, the type of living rate (\code{living.rate} argument), i.e. long-lived species (\code{living.rate = "long"}) or short-lived species (\code{living.rate = "short"}) must be quoted.
This argument is used to select the right formula from Niel & Lebreton (2005) for the calculation of lambdaMax (see \code{lambdaMax.L} for long-lived species and \code{lambdaMax.S} for short-lived species)
and, therefater Rmax.


The population size, as the previous terms, can also be provided either as point estimates (\code{pop.fixed}) or can be drawn from distributions (\code{pop.unif = TRUE}
with \code{min.pop} and \code{max.pop} or \code{pop.lognorm = TRUE} with \code{mean.pop} and \code{sd.pop}).

Theta is a shape parameter describing the form of density dependence. It can be provided as a point estimates (\code{theta.fixed}). If
the value of theta is unknown, it can be estimated from Rmax following the relationship described in Johnson \emph{et al.} (2012) :
\eqn{exp(1.129-1.824*Rmax+e)}, where e is either residuals of the underlying relationship set at their means (i.e. 0) (\code{estim.theta = "determinist"}) or
residuals drawn from normal distribution N(0, 0.942) (\code{estim.theta = "random"}.


The Sustainable Harvest Index (SHI) is calculated as the ratio harvest level. The harvest level can
be provided either as point estimates (\code{harvest.fixed}) or distributions (\code{harvest.unif = TRUE}
with \code{min.harvest} and \code{max.harvest} or \code{harvest.lognorm = TRUE} with \code{mean.harvest} and \code{sd.harvest}).
When harvest exceeds PTL (SHI>1), this suggests an overharvesting regime with regard to management objectives.
}
\examples{
PTL(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
estim.theta = c("random", "determinist"),
harvest.fixed = c(107802, 8447950),
Fobj = c(0.1, 0.3, 0.5))

}
\references{
Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Runge, M. C., Sauer, J. R., Avery, M. L., Blackwell, B. F., & Koneff, M. D. (2009). Assessing allowable take of migratory birds. The Journal of Wildlife Management, 73(4), 556-565.

Boitani, L., & Fuller, T. (Eds.). (2000). Research techniques in animal ecology: controversies and consequences. Columbia university press.
}
\seealso{
\code{\link{input.summary}}, \code{\link{output.summary}}, \code{\link{SHI.plot}}.
}
