% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\title{Maximum annual recruitment rate estimation}
\usage{
rmax(
  Nsim = 1,
  living.rate = NULL,
  surv.fixed = NULL,
  surv.norm = FALSE,
  mean.surv = NULL,
  sd.surv = NULL,
  mass.fixed = NULL,
  mass.norm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  type.p = NULL,
  type.e = NULL,
  alpha.fixed = NULL,
  alpha.unif = FALSE,
  min.alpha = NULL,
  max.alpha = NULL,
  alpha.norm = FALSE,
  mean.alpha = NULL,
  sd.alpha = NULL
)
}
\arguments{
\item{Nsim}{number, default 1 ;  number of simulations}

\item{living.rate}{character string ; living rate : either "\code{long}" or "\code{short}". No other character string allowed.}

\item{surv.fixed}{number ; fixed estimate of adult annual survival rate provided by the user. To be used instead of \code{mean.surv} when \code{surv.norm = FALSE}.}

\item{surv.norm}{surv.norm boolean, default FALSE. If \code{TRUE}, simulates adult annual survival rate from a normal distribution with mu = \code{mean.surv} and standard deviation = \code{sd.surv} (see below).}

\item{mean.surv}{mean.surv number ; arithmetic mean of adult annual survival rate. Only used if \code{surv.norm = TRUE}.}

\item{sd.surv}{sd.surv number ; standard deviation of adult annual survival rate. Only used if \code{surv.norm = TRUE}.}

\item{mass.fixed}{mass.fixed number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate from body mass.}

\item{mass.norm}{mass.norm boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{mean.mass number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.norm = TRUE}. #' @param sd.mass number ; standard deviation of body mass. Only used if \code{mass.norm = TRUE}.}

\item{sd.mass}{sd.mass number ; standard deviation of body mass. Only used if \code{mass.norm = TRUE}.}

\item{type.p}{type.p character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{type.e character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{alpha.fixed}{alpha.fixed number ; fixed estimate of age at first breeding (in years). To be used when \code{alpha.unif = FALSE} and \code{alpha.norm = FALSE}.}

\item{alpha.unif}{alpha.unif boolean, default \code{FALSE}. Simulates age at first breeding from a uniform distribution bounded by \code{min.alpha} and \code{max.alpha}.}

\item{min.alpha}{min.alpha number ; minimum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{max.alpha}{max.alpha number ; maximum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{alpha.norm}{alpha.norm boolean, default \code{FALSE}. If \code{TRUE}, simulates age at first breeding from a normal distribution with mu = \code{mean.alpha} and standard deviation = \code{sd.alpha}.}

\item{mean.alpha}{mean.alpha number ; arithmetic mean of age at first breeding. To be used only used if \code{alpha.norm = TRUE}.}

\item{sd.alpha}{sd.alpha number ; age at first breeding standard deviation. Only used if \code{alpha.norm = TRUE}.}
}
\value{
A dataframe
}
\description{
For internal use only.
}
\references{
Dillingham, P. W., & Fletcher, D. (2008). Estimating the ability of birds to sustain additional human-caused mortalities using a simple decision rule and allometric relationships. Biological Conservation, 141(7), 1783-1792.

Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Niel, C., & LEBRETON, J. D. (2005). Using demographic invariants to detect overharvested bird populations from incomplete data. Conservation Biology, 19(3), 826-835.
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
