% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHI.plot.R
\name{SHI.plot}
\alias{SHI.plot}
\title{Sustainability harvest index visualizations from PEG and PTL functions}
\usage{
SHI.plot(
  data = NULL,
  Sp = NULL,
  IDSp = TRUE,
  NameSp = NULL,
  title = "Distribution of sustainability harvest index",
  title.x = "Sustainability harvest index",
  title.y = "Number of simulations",
  color1 = "#00BFC4",
  color2 = "#F8766D"
)
}
\arguments{
\item{data}{output dataframe from \code{PEG} and \code{PTL} functions. Other inputs will give an error.}

\item{Sp}{number ; the identification number assigned to each studied species}

\item{IDSp}{boolean ; default \code{TRUE}. If \code{TRUE}, print the identification number assigned to each studied species. If \code{FALSE}, doesn't print the identification number assigned to the studied species.}

\item{NameSp}{character string ; the name of the studies species. Add the names of the studies species to the main title of the plot. Example : c("Species A", "Species B")}

\item{title}{character string ; default "Distribution of sustainability harvest index" ; main title}

\item{title.x}{character string ; default "Sustainability harvest index" ; title x label}

\item{title.y}{character string ; default "Simulations number" ; title y label}

\item{color1}{character string ; default green (#008000)}

\item{color2}{character string ; default red (#B40000)}
}
\value{
visualizations
}
\description{
Create visualizations of sustainability harvest indices calculated from \code{PEG} or \code{PTL} functions
}
\examples{
df <- PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

SHI.plot(df)
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
