% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv}
\alias{surv}
\title{Survival estimation}
\usage{
surv(
  Nsim = 1,
  NSp = 1,
  mass.fixed = NULL,
  mass.lognorm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  alpha = NULL,
  type.p = NULL,
  type.e = NULL,
  surv.j = 1
)
}
\arguments{
\item{Nsim}{number ; number of simulations.}

\item{NSp}{number ; number of species for which calculations are performed.}

\item{mass.fixed}{number ; point estimate of body mass (in kilograms) used to estimate annual survival rate of adults.}

\item{mass.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a log-normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.lognorm = TRUE}.}

\item{sd.mass}{number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{alpha}{number ; age at the first reproduction}

\item{type.p}{character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{surv.j}{number, default 1 ; point estimate average annual survival for birds aged 1 to alpha. The value can be provided by the user to estimate adult survival rate from body mass under certain circumstances (see 'Details).}
}
\value{
A dataframe with the following information for each of the n simulations :
\itemize{
  \item \code{Sim} number of the simulation.
  \item \code{mass} body mass (in kilograms).
  \item \code{e} residuals
  \item \code{p}
  \item \code{survival} estimate of annual survival rate of adult under ideal conditions.
}
}
\description{
\code{surv} is a function used to estimate (maximal) annual survival rate of adults under ideal conditions, from body mass (in kilograms), see 'Details'.
}
\details{
Adult survival rate can be estimated from body mass following the relationship described in Johnson \emph{et al.} (2012) : \eqn{(p/surv.j)^(1/(exp(3.22+0.24*log(body mass)+e)-alpha))}.
IBody mass can be provided as either a point estimate (\code{mass.fixed}) or drawn from
a log-normal distribution (\code{mass.lognorm = TRUE}, \code{mean.mass} and \code{sd.mass} arguments). The parameter p can be either drawn from a beta distribution
B(3.34, 101.24) (\code{type.p = "random"}) or provided as a point estimate equal to 0.032 (i.e. the mean of the beta distribution is given by a/(a + b) with a = 3.34 and b = 101.24, Boitani & Fuller 2000) (\code{type.p = "determinist"})
The parameter e (residuals) can be either set to zero (\code{type.e = "determinist"}) or residuals drawn of a normal distribution N(0, 0.087) (\code{type.e = "random"}). See Johnson \emph{et al.} (2012).


By default, the argument \code{surv.j} is silent and masked to users (the default value for \code{surv.j} is 1, as such \eqn{p/surv.j = p/1 = p} ; see original equation in
Johnson \emph{et al.}, 2012). This is correct for any alpha if, and only if, juvenile survival < adult survival solely for birds aged <1 year.  This is a reasonable
approximation especially in passerines and some other medium-sized bird species. For species with delayed sexual maturity and when juvenile survival < adult
survival for birds aged 1 to alpha, the default \code{surv} function returns biased estimates unless the users specify a value for annual survival for birds aged 1
to alpha. This is done by using the argument \code{surv.j} which is then used to estimate  \eqn{p = p/surv.j}.
}
\references{
Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Boitani, L., & Fuller, T. (Eds.). (2000). Research techniques in animal ecology: controversies and consequences. Columbia university press.
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}
}
