% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.summary.R
\name{input.summary}
\alias{input.summary}
\title{Summary of input parameters used by \code{PEG} and \code{PTL} functions}
\usage{
input.summary(data)
}
\arguments{
\item{data}{dataframe from \code{PEG} and \code{PTL} functions input data. Other inputs will give an error.}
}
\value{
A dataframe with parameters and dispersion statistics in rows. Species(SP), safety factor (Fs) or objective factor (Fobj) are in column :
In rows :
\itemize{
  \item \code{SP} the species
}
In columns :
\itemize{
  \item \code{Fs} safety factor
  \item \code{Fobj} objective factor
  \item \code{mean.surv} arithmetic mean of adult annual survival rate
  \item \code{med.surv} median of adult annual survival rate
  \item \code{percent2.5.surv} 2.5th pencentile of adult annual survival rate
  \item \code{percent97.5.surv} 97.5th percentile of adult annual survival rate
  \item \code{mean.alpha} arithmetic mean of age at first breeding
  \item \code{med.alpha} median of age at first breeding
  \item \code{percent2.5.alpha} 2.5th pencentile of age at first breeding
  \item \code{percent97.5.alpha} 97.5th percentile of age at first breeding
  \item \code{mean.lambdaMax} arithmetic mean of maximum annual growth rate
  \item \code{med.lambdaMax} median of maximum annual growth rate
  \item \code{percent2.5.lambdaMax} 2.5th pencentile of maximum annual growth rate
  \item \code{percent97.5.lambdaMax} 97.5th percentile of maximum annual growth rate
  \item \code{mean.Rmax} arithmetic mean of maximum annual recruitment rate
  \item \code{med.Rmax} median of maximum annual recruitment rate
  \item \code{percent2.5.Rmax} 2.5% pencentile of maximum annual recruitment rate
  \item \code{percent97.5.Rmax} 97.5% percentile of maximum annual recruitment rate
}
}
\description{
The function returns descriptive statistics (mean, median, se and 95 \% confidence intervals) for input parameters used by PEG or PTL functions : annual survival rate of adults (surv), age at first breeding (alpha), maximum annual growth rate (lambdaMax) and maximum annual recruitment rate (Rmax). Descriptive statistics are drawn from \emph{n} simulations called within the \code{PEG} and \code{PTL} functions. When one or several parameters are fixed by the user, the corresponding value is reported for both mean and median, without any further measure of dispersion (i.e. se and 95 \% CI are set zero). Statistics are reported in columns.
}
\examples{
df <- PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

input.summary(df)

}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
