% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaMax.L.R
\name{lambdaMax.L}
\alias{lambdaMax.L}
\title{Maximum annual growth rate for long-lived species}
\usage{
lambdaMax.L(alpha, survival)
}
\arguments{
\item{alpha}{number ; age at first breeding}

\item{survival}{number ; annual survival rate of the adult age class}
}
\value{
A number
}
\description{
For internal use only. The function estimates maximum annual growth rate for long-lived species using the demographic invariants approach (Niel & Lebreton 2005).
}
\references{
Niel, C., & Lebreton, J. D. (2005). Using demographic invariants to detect overharvested bird populations from incomplete data. Conservation Biology, 19(3), 826-835.
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}
}
