% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL.R
\name{PTL}
\alias{PTL}
\title{Prescribed Take Level estimation}
\usage{
PTL(
  full.option = FALSE,
  Nsim = 1,
  NSp = NULL,
  Rmax.fixed = NULL,
  Rmax.lognorm = FALSE,
  mean.Rmax = NULL,
  sd.Rmax = NULL,
  lambdaMax.fixed = NULL,
  lambdaMax.lognorm = FALSE,
  mean.lambdaMax = NULL,
  sd.lambdaMax = NULL,
  living.rate = NULL,
  surv.fixed = NULL,
  surv.beta = FALSE,
  mean.surv = NULL,
  sd.surv = NULL,
  mass.fixed = NULL,
  mass.lognorm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  type.p = NULL,
  type.e = NULL,
  alpha.fixed = NULL,
  alpha.unif = FALSE,
  min.alpha = NULL,
  max.alpha = NULL,
  alpha.lognorm = FALSE,
  mean.alpha = NULL,
  sd.alpha = NULL,
  Fobj = NULL,
  harvest.fixed = NULL,
  harvest.unif = FALSE,
  min.harvest = NULL,
  max.harvest = NULL,
  harvest.lognorm = FALSE,
  mean.harvest = NULL,
  sd.harvest = NULL,
  pop.fixed = NULL,
  pop.unif = FALSE,
  min.pop = NULL,
  max.pop = NULL,
  pop.lognorm = FALSE,
  mean.pop = NULL,
  sd.pop = NULL,
  theta.fixed = NULL,
  estim.theta = NULL
)
}
\arguments{
\item{full.option}{boolean, default \code{FALSE}. If \code{TRUE}, provides all parameter values necessary for calculations.}

\item{Nsim}{number, default 1 ;  number of simulations.}

\item{NSp}{number ; number of species for which calculations are done.}

\item{Rmax.fixed}{number ; fixed estimate of maximum annual growth rate provided by the user. To be used instead of \code{mean.Rmax} when \code{Rmax.lognorm = FALSE}.}

\item{Rmax.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates Rmax for a normal distribution with mu = \code{mean.Rmax} and standard deviation = \code{sd.Rmax}.}

\item{mean.Rmax}{number ; marithmetic mean of maximum annual growth rate.}

\item{sd.Rmax}{number ; standard deviation of maximum annual growth rate.}

\item{lambdaMax.fixed}{number ; fixed estimate of maximum multiplicative growth rate provided by the user and used to calculate Rmax as \eqn{lambdaMax – 1}.}

\item{lambdaMax.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates maximum mulitplicative growth rate from a normal distribution with mu = \code{mean.lambdaMax} and standard deviation = \code{sd.lambdaMax}.}

\item{mean.lambdaMax}{a number ; arithmetic mean of maximum multiplicative growth rate. To be used only used if \code{lambdaMax.lognorm = TRUE}.}

\item{sd.lambdaMax}{a number ; standard deviation of maximum multiplicative growth rate. To be used only if \code{lambdaMax.lognorm = TRUE}.}

\item{living.rate}{character string ; living rate : either \code{"long"} or \code{"short"}, see 'Details'. No ohter arguments allowed.}

\item{surv.fixed}{number ; fixed estimate of adult annual survival rate provided by the user. To be used instead of mean.surv when \code{surv.beta = FALSE}.}

\item{surv.beta}{boolean, default \code{FALSE}. If \code{TRUE}, simulates adult annual survival rate from a bate distribution with mu = \code{mean.surv} and standard deviation = \code{sd.surv} (see below).}

\item{mean.surv}{number ; arithmetic mean of adult annual survival rate. To be used only if \code{surv.beta = TRUE}.}

\item{sd.surv}{number ; standard deviation of adult annual survival rate. To be used only if \code{surv.beta = TRUE}.}

\item{mass.fixed}{number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate from body mass.}

\item{mass.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.lognorm = TRUE}.}

\item{sd.mass}{number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{type.p}{character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{alpha.fixed}{number ; number ; fixed estimate of age at first breeding (in years). To be used when \code{alpha.unif = FALSE} and \code{alpha.lognorm = FALSE}.}

\item{alpha.unif}{boolean, default \code{FALSE}. Simulates age at first breeding from a uniform distribution bounded by \code{min.alpha} and \code{max.alpha}.}

\item{min.alpha}{number ; minimum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{max.alpha}{number ; maximum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{alpha.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates age at first breeding from a normal distribution with mu = \code{mean.alpha} and standard deviation = \code{sd.alpha}.}

\item{mean.alpha}{number ; arithmetic mean of age at first breeding. To be used only used if \code{alpha.lognorm = TRUE}.}

\item{sd.alpha}{number ; standard deviation of age at first breeding. To be used only if \code{alpha.lognorm = TRUE}.}

\item{Fobj}{number ; objective factor, see 'Details'.}

\item{harvest.fixed}{number ; number of harvested individuals.}

\item{harvest.unif}{boolean ; default \code{FALSE}. If \code{TRUE}, simulates harvest level from a uniform distribution bounded by \code{min.harvest} and \code{max.harvest}.}

\item{min.harvest}{number ; minimum number of harvested individuals. To be used only if \code{harvest.unif = TRUE}.}

\item{max.harvest}{number ; maximum number of individuals harvested. To be used only if \code{harvest.unif = TRUE}.}

\item{harvest.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates harvest level from a normal distribution with mu =  \code{mean.harvest} and standard deviation = \code{sd.harvest}.}

\item{mean.harvest}{number ; arithmetic mean of the number of harvested individuals. To be used only if \code{harvest.lognorm = TRUE}.}

\item{sd.harvest}{number ; standard deviation of the number of harvested individuals. To be used only if \code{harvest.lognorm = TRUE}.}

\item{pop.fixed}{number ;  fixed estimate of population size (in number of individuals). To be used when \code{pop.unif = FALSE} and \code{pop.lognorm = FALSE}.}

\item{pop.unif}{boolean, default \code{FALSE}. If \code{TRUE}, simulates a population size from a uniform distribution bounded by \code{min.pop} and \code{max.pop}.}

\item{min.pop}{number ; minimum population size. To be used only if \code{pop.unif = TRUE}.}

\item{max.pop}{number ; maximum population size. To be only used if \code{pop.unif = TRUE}.}

\item{pop.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates a population size from a normal distribution with mu = \code{mean.pop} and standard deviation = \code{sd.pop}.}

\item{mean.pop}{number ; arithmetic mean of population size. To be used only if \code{pop.lognorm = TRUE}.}

\item{sd.pop}{number ; standard deviation of population size. To be used only if \code{pop.lognorm = TRUE}.}

\item{theta.fixed}{number ; parameter for density-dependance.}

\item{estim.theta}{character string ; default "\code{determinist}". The value of the parameter theta can be either "\code{determinist}" or "\code{random}", see 'Details'. No other arguments allowed.}
}
\value{
A dataframe with the following informations for each the n simulations in addition to the information provided in the function's arguments :
If the argument \code{full.option} is \code{FALSE}, it provides a simple dataframe :
\itemize{
  \item \code{SP} the identification number assigned to the studied species
  \item \code{Sim} the index of the simulation.
  \item \code{Fs} the value of the safety factor.
  \item \code{Rmax} the maximal annual recruitment rate (lambda max – 1).
  \item \code{popsize} the population size (in number of individuals).
  \item \code{harvest}  the number of individuals harvested each year.
  \item \code{PTL} the Prescribed take level expressed in  individuals number.
  \item \code{SHI} Sustainability harvest index (from 0 to +∞).
}
If the argument \code{full.option} is \code{TRUE}, it can provide the following informations in addition to the these within the simple dataframe :
\itemize{
  \item \code{e} a value extracted from a normal distribution (0, 0.087). This parameter appears when the adult annual survival rate is estimated from the body mass.
  \item \code{p} a value extracted from a beta distribution (3.34,101.24). This parameter appears when the adult annual survival rate is estimated from the body mass.
  \item \code{e.density} a value extracted from a normal distribution (0, 0.942). This parameter appears when the value of the parameter for density-dependance is unknown and must be estimated.
}
}
\description{
\code{PTL} is a function used to calculate the prescribed take level of a species, see 'Details'.
}
\details{
Calculates the Prescribed Take Level (PTL) integrating a density-dependance parameter and Sustainability Harvest Index (SHI).

PTL is the maximum number of animals which may be removed from a population while allowing it to
reach or maintain equilibrium to a certain level of its carrying capacity. The PTL formula is
\eqn{Fobj.(Rmax.theta)/(1+theta)} for a breeding population.


Fobj (\code{Fobj} argument) is a factor that reflects management objectives of the decision-maker. It is related to
the desired long-term population size relative to the carrying capacity (Runge et al. 2009). For example,
a factor of 0.25 means that 25% of the maximum sustainable yield is taken.


Rmax is the maximum annual recruitment rate estimation. This is equivalent to the maximum annual population
growth rate, lambdaMax, under optimal growth conditions. Rmax and lambdaMax can be provided by either a point estimates (\code{Rmax.fixed}, \code{lambdaMax.fixed}) or a normal distribution (\code{Rmax.lognorm = TRUE}
with \code{mean.Rmax} and \code{sd.Rmax}, \code{lambdaMax = TRUE} with \code{mean.lambdaMax} and \code{sd.lambdaMax}). When Rmax (or lambdaMax) is unknown from the user,
the value is calculated using the demographic invariants
approach (Niel & Lebreton 2005). This approach calculates \eqn{Rmax = lambda max – 1} on the basis of 2 parameters:
adult survival rate and age at first breeding. When adult survival and age at first reproduction are known
from the user, they can be provided by either point estimates (\code{surv.fixed}, \code{alpha.fixed} arguments) or distributions (\code{surv.beta = TRUE} with \code{mean.surv} and \code{sd.surv},
\code{alpha.unif = TRUE} with \code{min.alpha} and \code{max.alpha}, \code{alpha.lognorm = TRUE} with \code{mean.alpha} and \code{sd.alpha} arguments). Also,
survival rate can be estimated from body mass following the relationship described in Johnson et al. (2012) : \eqn{p^(1/(exp(3.22+0.24*log(body mass)+e)-1))}.
In this case, body mass can be provided as either a point estimate (\code{mass.fixed})
or a normal distribution (\code{mass.lognorm = TRUE}, \code{mean.mass} and \code{sd.mass} arguments). The parameter p can be provided by either from a beta distribution
B(3.34, 101.24) (\code{type.p = "random"}) or a point estimate equal to 0.032 (i.e. the mean of the beta distribution is given by a/(a + b) with a = 3.34 and b = 101.24, Boitani & Fuller 2000) (\code{type.p = "determinist"})
The parameter e can also be either null (\code{type.e = "determinist"}) or residuals from a normal distribution N(0, 0.087) (\code{type.e = "random"}).


In all cases,  the type of living rate (\code{living.rate} argument), i.e. long-lived species (\code{living.rate = "long"}) or short-lived species (\code{living.rate = "short"}) must be quoted. Long-lived species,
unlike short-lived species, are those that do not have a strategy to maximize reproduction, i.e. those with high adult survival, low brood size,
and high first-generation age. This argument is used to select the right formula from Niel & Lebreton (2005) for the calculation of lambdaMax
and, therefater Rmax.


The population size, as the previous terms, can also be provided by either a point estimates (\code{pop.fixed}) or distributions (\code{pop.unif = TRUE}
with \code{min.pop} and \code{max.pop} or \code{pop.lognorm = TRUE} with \code{mean.pop} and \code{sd.pop}).

Theta is a shape parameter describing the form of density dependence. It can be provided by either a point estimates (\code{theta.fixed}). If
the value of theta is unknown, it can be estimated from Rmax following the relationship described in Johnson et al (2012) :
\eqn{exp(1.129-1.824*Rmax+e)}, where e is either residuals of the underlying relationship at their means (i.e. 0) (\code{estim.theta = "determinist"}) or
residuals from normal distribution N(0, 0.942) (\code{estim.theta = "random"}.


The Sustainability Harvest Index or SHI. It is the harvest level expressed as a proportion relative to PTL: \eqn{harvest level/PTL}. The harvest level can
be provided by a point estimates (\code{harvest.fixed}) or distributions (\code{harvest.unif = TRUE}
with \code{min.harvest} and \code{max.harvest} or \code{harvest.lognorm = TRUE} with \code{mean.harvest} and \code{sd.harvest}).
When harvest exceeds PTL, this suggests an overexploitation regime according a fixed management objective.


The function PEG provides a great flexibility of use according the known arguments and analysis that the user wants realize.
Fixed demographic parameters or different distribution types can be integrated in the function. This later approach enables uncertainty
to be taken into account.


For more details for the function use see 'README' documentation.
}
\examples{
PTL(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
estim.theta = c("random", "determinist"),
harvest.fixed = c(107802, 8447950),
Fobj = c(0.1, 0.3, 0.5))

}
\references{
Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Runge, M. C., Sauer, J. R., Avery, M. L., Blackwell, B. F., & Koneff, M. D. (2009). Assessing allowable take of migratory birds. The Journal of Wildlife Management, 73(4), 556-565.

Boitani, L., & Fuller, T. (Eds.). (2000). Research techniques in animal ecology: controversies and consequences. Columbia university press.
}
\seealso{
\code{\link{input.summary}}, \code{\link{output.summary}}, \code{\link{SHI.plot}}.
}
