% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\title{Maximum annual recruitment rate estimation}
\usage{
rmax(
  Nsim = 1,
  living.rate = NULL,
  surv.fixed = NULL,
  surv.beta = FALSE,
  mean.surv = NULL,
  sd.surv = NULL,
  mass.fixed = NULL,
  mass.lognorm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  type.p = NULL,
  type.e = NULL,
  surv.j = 1,
  alpha.fixed = NULL,
  alpha.unif = FALSE,
  min.alpha = NULL,
  max.alpha = NULL,
  alpha.lognorm = FALSE,
  mean.alpha = NULL,
  sd.alpha = NULL
)
}
\arguments{
\item{Nsim}{number, default 1 ;  number of simulations}

\item{living.rate}{character string ; living rate : either "\code{long}" or "\code{short}". No other character string allowed.}

\item{surv.fixed}{number ; fixed estimate of (annual) adult survival rate provided by the user. To be used instead of \code{mean.surv} when \code{surv.beta = FALSE}.}

\item{surv.beta}{surv.beta boolean, default FALSE. If \code{TRUE}, simulates (annual) adult survival rate from a beta distribution with mu = \code{mean.surv} and standard deviation = \code{sd.surv} (see below).}

\item{mean.surv}{mean.surv number ; arithmetic mean of (annual) adult survival rate. Only used if \code{surv.beta = TRUE}.}

\item{sd.surv}{sd.surv number ; standard deviation of adult annual survival rate. Only used if \code{surv.beta = TRUE}.}

\item{mass.fixed}{mass.fixed number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate from body mass.}

\item{mass.lognorm}{mass.lognorm boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{mean.mass number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.lognorm = TRUE}. #' @param sd.mass number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{sd.mass}{sd.mass number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{type.p}{type.p character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{type.e character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{surv.j}{surv.j number, default 1 ; point estimate of average annual survival for birds aged 1 to alpha. The value can be provided by the user. Used to estimate adult survival rate from body mass under certain circumstances (see Details).}

\item{alpha.fixed}{alpha.fixed number ; fixed estimate of age at first breeding (in years). To be used when \code{alpha.unif = FALSE} and \code{alpha.lognorm = FALSE}.}

\item{alpha.unif}{alpha.unif boolean, default \code{FALSE}. Simulates age at first breeding from a uniform distribution bounded by \code{min.alpha} and \code{max.alpha}.}

\item{min.alpha}{min.alpha number ; minimum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{max.alpha}{max.alpha number ; maximum age at first breeding. To be used only if \code{alpha.unif = TRUE}.}

\item{alpha.lognorm}{alpha.lognorm boolean, default \code{FALSE}. If \code{TRUE}, simulates age at first breeding from a normal distribution with mu = \code{mean.alpha} and standard deviation = \code{sd.alpha}.}

\item{mean.alpha}{mean.alpha number ; arithmetic mean of age at first breeding. To be used only used if \code{alpha.lognorm = TRUE}.}

\item{sd.alpha}{sd.alpha number ; age at first breeding standard deviation. Only used if \code{alpha.lognorm = TRUE}.}
}
\value{
A dataframe
}
\description{
\code{rmax} is a function used to estimate the maximum annual recruitment rate, see 'Details'.
}
\details{
When the maximum annual recruitment rate, Rmax, is unknown,
the value can be estimated using the Demographic Invariants Method (DIM, Niel & Lebreton, 2005). This approach calculates \eqn{Rmax = lambdaMax – 1} (Dillingham & Fletcher, 2008) on the basis of 2 parameters:
annual survival rate of adult under ideal condition (maximal survival, \code{surv} arguments) and age at first breeding \code{alpha} arguments). When adult survival and age at first breeding are known,
users can be provide corresponding values either as point estimates (\code{surv.fixed}, \code{alpha.fixed} arguments) or as distributions (\code{surv.beta = TRUE} with \code{mean.surv} and \code{sd.surv},
\code{alpha.unif = TRUE} with \code{min.alpha} and \code{max.alpha}, \code{alpha.lognorm = TRUE} with \code{mean.alpha} and \code{sd.alpha} arguments). Alternatively,
adult survival rate can be estimated from body mass following the relationship described in Johnson \emph{et al.} (2012) : \eqn{(p/surv.j)^(1/(exp(3.22+0.24*log(body mass)+e)-alpha))} (see \code{surv} function).
Body mass can be provide either as a point estimate (\code{mass.fixed})
or drawn from a log-normal distribution (\code{mass.lognorm = TRUE}, \code{mean.mass} and \code{sd.mass} arguments). The parameter p can be drawn either from a beta distribution
B(3.34, 101.24) (\code{type.p = "random"}) or a point estimate equal to 0.032 (i.e. the mean of the beta distribution is given by a/(a + b) with a = 3.34 and b = 101.24, Boitani & Fuller 2000) (\code{type.p = "determinist"})
The parameter e (residuals) can be either set to zero (\code{type.e = "determinist"}) or residuals from a normal distribution N(0, 0.087) (\code{type.e = "random"}). See equations in Johnson \emph{et al.} (2012).


By default, the argument \code{surv.j} is silent and masked to users (the default value for \code{surv.j} is 1, as such \eqn{p/surv.j = p/1 = p} ; see original equation in
Johnson \emph{et al.}, 2012). This is correct for any alpha if, and only if, juvenile survival < adult survival solely for birds aged <1 year.  This is a
reasonable approximation especially in passerines and some other medium-sized bird species. For species with delayed sexual maturity and when juvenile
survival < adult survival for birds aged 1 to alpha, the default \code{surv} function returns biased estimates unless the users specify a value for annual survival
for birds aged 1 to alpha. This is done by using the argument \code{surv.j} which is then used to estimate \eqn{p = p/surv.j}.
}
\references{
Dillingham, P. W., & Fletcher, D. (2008). Estimating the ability of birds to sustain additional human-caused mortalities using a simple decision rule and allometric relationships. Biological Conservation, 141(7), 1783-1792.

Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Niel, C., & LEBRETON, J. D. (2005). Using demographic invariants to detect overharvested bird populations from incomplete data. Conservation Biology, 19(3), 826-835.
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
