% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHI.plot.R
\name{SHI.plot}
\alias{SHI.plot}
\title{Sustainability harvest index visualizations from PEG and PTL functions}
\usage{
SHI.plot(
  data = NULL,
  Sp = NULL,
  IDSp = TRUE,
  NameSp = NULL,
  title = "Distribution of sustainability harvest index",
  title.x = "Sustainability harvest index",
  title.y = "Number of simulations",
  color1 = "#00BFC4",
  color2 = "#F8766D"
)
}
\arguments{
\item{data}{dataframe from \code{PEG} and \code{PTL} functions output. Other inputs will give an error.}

\item{Sp}{number ; the identification number assigned to the studied species}

\item{IDSp}{boolean ; default \code{TRUE}. If \code{TRUE}, prints the the identification number assigned to the studied species. If \code{FALSE}, doesn't print the the identification number assigned to the studied species.}

\item{NameSp}{character string ; the name of the studies species. Adds the names of the studies species to the title of the graphic.}

\item{title}{character string ; default "Distribution of sustainability harvest index" ; title plot}

\item{title.x}{character string ; default "Sustainability harvest index" ; title axis x}

\item{title.y}{character string ; default "Simulations number" ; title axis y}

\item{color1}{character string ; default green (#008000) ; the first color that the user wants for the graphic}

\item{color2}{character string ; default red (#B40000) ; the second color that the user wants for the graphic}
}
\value{
data visualisations
}
\description{
Create data visualizations of sustainability harvest index from \code{PEG} and \code{PTL} functions
}
\examples{
df <- PEG(Nsim = 20, NSp = 2, living.rate = c("long", "short"),
surv.fixed = c(0.8, 0.65),
alpha.unif = TRUE, min.alpha = c(2, 1), max.alpha = c(3, 2),
pop.fixed = c(3605244, 55805898),
harvest.fixed = c(107802, 8447950),
Fs = c(0.1, 0.3, 0.5))

SHI.plot(df)
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}.
}
