% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv}
\alias{surv}
\title{Survival estimation}
\usage{
surv(
  Nsim = 1,
  NSp = 1,
  mass.fixed = NULL,
  mass.lognorm = FALSE,
  mean.mass = NULL,
  sd.mass = NULL,
  alpha = NULL,
  type.p = NULL,
  type.e = NULL
)
}
\arguments{
\item{Nsim}{number ; number of simulations.}

\item{NSp}{number ; number of species for which calculations are done.}

\item{mass.fixed}{number ; fixed estimate of body mass (in kilograms) used to estimate adult survival rate from body mass.}

\item{mass.lognorm}{boolean, default \code{FALSE}. If \code{TRUE}, simulates body mass from a normal distribution with mu = \code{mean.mass} and standard deviation = \code{sd.mass}.}

\item{mean.mass}{number ; arithmetic mean of body mass (in kilograms). Only used if \code{mass.lognorm = TRUE}.}

\item{sd.mass}{number ; standard deviation of body mass. Only used if \code{mass.lognorm = TRUE}.}

\item{alpha}{number ; age at the first reproduction}

\item{type.p}{character string ; the value of the parameter p can be either "\code{determinist}" or "\code{random}". No other character string allowed.}

\item{type.e}{character string ; the value of the parameter e can be either "\code{determinist}" or "\code{random}". No other character string allowed.}
}
\value{
A dataframe with the following information for each of the n simulations :
\itemize{
  \item \code{Sim} number of the simulation.
  \item \code{mass} body mass (in kilograms).
  \item \code{e}
  \item \code{p}
  \item \code{survival} estimate of adult annual survival rate.
}
}
\description{
For internal use only. Survival estimation based on body mass.
}
\references{
Johnson, F. A., Walters, M. A., & Boomer, G. S. (2012). Allowable levels of take for the trade in Nearctic songbirds. Ecological Applications, 22(4), 1114-1130.

Boitani, L., & Fuller, T. (Eds.). (2000). Research techniques in animal ecology: controversies and consequences. Columbia university press.
}
\seealso{
\code{\link{PEG}}, \code{\link{PTL}}
}
